Better sandbox lifecycle management (#95)
* Sandbox start/stop owner w/ configurable delay

Previously, the test process itself was the made owner of the sandbox
process, which meant the sandbox would be terminated when the test
terminated.

Unlike some other kinds of tests, playwright tests are not linked to the
LiveViews and other processes that are being tested, so those processes
don't terminate with the test and may go on trying to use the sandbox
connection after the test terminates. If they do, they will raise
ownership errors.

Switch to `Ecto.Adapters.SQL.Sandbox.html.start_owner!/2`, creating a
separate process to own the sandbox connection, which is shut down after
the test process terminates using `on_exit/1`. This reduces ownership
errors that may occur after the test terminates.

By default, this is done with no delay, but `sandbox_shutdown_delay` can
be configured or specified via tag, which will wait that long before
terminating the sandbox during `on_exit/1`.